Update source files to use moved exception module
